7. final adaptation of SQLite to the project

+++ app/src/main/java/com/example/navcompro/screens/main/toobs/dashboard/BoxFragment.kt
@@ -46,7 +46,7 @@ class BoxFragment : Fragment(R.layout.fragment_box) {
         }
     }
 
-    private fun getBoxId(): Int = args.boxId
+    private fun getBoxId(): Long = args.boxId
         //"Extract box id from arguments here")
